Handle missing pids

+++ files/usr/local/bin/mgr/clean_zombie.lua
@@ -19,24 +19,27 @@ function clean()
         do
             local zombie = false
             local ppid = nil
-            for k, line in ipairs(read_all("/proc/" .. pid .. "/status"):splitNewLine())
-            do
-                -- Look for a zombie
-                local m = string.match(line, "State:%s[ZT]")
-                if m then
-                    zombie = true
-                end
-                if zombie then
-                    m = string.match(line, "PPid:%s([0-9]*)")
+            local all = read_all("/proc/" .. pid .. "/status")
+            if all then
+                for k, line in ipairs(all:splitNewLine())
+                do
+                    -- Look for a zombie
+                    local m = string.match(line, "State:%s[ZT]")
                     if m then
-                        ppid = m
-                        break
+                        zombie = true
+                    end
+                    if zombie then
+                        m = string.match(line, "PPid:%s([0-9]*)")
+                        if m then
+                            ppid = m
+                            break
+                        end
                     end
                 end
-            end
-            if ppid and ppid ~= 1 then
-                log:write("Killed " .. ppid)
-                nixio.kill(ppid, 9)
+                if ppid and ppid ~= 1 then
+                    log:write("Killed " .. ppid)
+                    nixio.kill(ppid, 9)
+                end
             end
         end
     end
